+++ docs/.vuepress/head.js
@@ -1,0 +1,3 @@
+module.exports = [
+    ['link', { rel: 'icon', href: 'https://github.com/slimkit/plus/blob/master/public/favicon.ico?raw=true' }]
+];