+++ packages/slimkit-plus-news/src/API2/Controllers/RankController.php
@@ -62,7 +62,7 @@ class RankController extends Controller
         ->where('created_at', '>', $date)
         ->where('audit_status', 0)
         ->with(['user' => function ($query) {
-            return $query->select('id', 'name', 'sex')->withTrashed();
+            return $query->select('id', 'name', 'sex', 'avatar')->withTrashed();
         }])
         ->groupBy('user_id')
         ->orderBy('count', 'desc')